完成了所有Service层的测试。添加了部分Mapper

+++ smart-retail-inventory/src/main/java/com/retail/inventory/service/SysRoleService.java
@@ -17,7 +17,7 @@ public interface SysRoleService {
      * @param role
      * @return
      */
-    int addRole(SysRole role);
+    Long addRole(SysRole role);
     /**
      * 更新角色
      * @param role
